poller/requirements.txt
Flask-SQLAlchemy==2.3.2
psycopg2==2.8.6
python-nmap==0.6.1
arrow==0.17.0
python-owasp-zap-v2.4==0.0.18
